Added type conversion.

+++ include/arctk/geom/shape/mesh.hpp
@@ -253,7 +253,7 @@ namespace arc //! arctk namespace
                     }
                 }
 
-                return (edges.size());
+                return (static_cast<unsigned int>(edges.size()));
             }
 
             /**
